+++ docs/fr_FR/changelog.md
@@ -4,4 +4,10 @@
 >
 >S'il n'y a pas d'information sur la mise à jour, c'est que celle-ci concerne uniquement de la mise à jour de documentation, de traduction ou de texte.
 
-Version stable non publiée : beta uniquement
+# 05/02/2025 / 1.3
+ - Refonte gestion caractères unicode reçus et envoyés avec l'écran
+ - Correction bug qui empêchait les Info xxx.text ou xxx.val de s'afficher (commande Info)
+ - Correction de la configuration par défaut lors de l'installation du plugin
+
+# xx/xx/xx / < 1.3
+ - Voir changelog beta